♻️ move libs folder & rename cfg.h
