+++ impl-c/include/libs/string.h
