+++ impl-c/proc/exec.c
@@ -10,7 +10,7 @@
 #include "timer.h"
 #include "uart.h"
 
-#include "cfg.h"
+#include "config.h"
 #include "log.h"
 
 #ifdef CFG_LOG_PROC_EXEC
