+++ impl-c/proc/sched.c
@@ -5,7 +5,7 @@
 #include "mm.h"
 #include "uart.h"
 
-#include "cfg.h"
+#include "config.h"
 #include "log.h"
 
 #ifdef CFG_LOG_PROC_SCHED
