+++ impl-c/proc/task.c
@@ -13,7 +13,7 @@
 #include "timer.h"
 #include "uart.h"
 
-#include "cfg.h"
+#include "config.h"
 #include "log.h"
 
 #include <stdint.h>
